requirements.txt
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///D:/bld/aiohttp_1767524610135/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1751626463503/work
altair==6.0.0
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
attrs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_attrs_1764875182/work
backoff==2.2.1
backports.zstd @ file:///D:/bld/bld/rattler-build_backports.zstd_1767045012/work
bcrypt==5.0.0
blinker==1.9.0
Brotli @ file:///D:/bld/brotli-split_1764017006836/work
build==1.3.0
cachetools==6.2.4
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1767500808759/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
chromadb==1.4.0
click==8.3.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.2
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1690101045195/work
distro==1.9.0
docstring_parser==0.17.0
duckdb==1.4.3
durationpy==0.10
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1767377229747/work
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist @ file:///D:/bld/frozenlist_1752167156313/work
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.45.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet @ file:///D:/bld/bld/rattler-build_hf-xet_1763772668/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1761225853865/work
humanfriendly==10.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jinja2_1764517220/work
jiter==0.12.0
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1765794027227/work
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.2.0
langchain-anthropic==1.3.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
markdown-it-py==4.0.0
MarkupSafe @ file:///D:/bld/markupsafe_1759055179018/work
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mmh3==5.2.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
multidict @ file:///D:/bld/multidict_1765460294251/work
multiprocess @ file:///D:/bld/multiprocess_1695458970146/work
mypy_extensions==1.1.0
narwhals==2.14.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1765215107/work
numpy @ file:///D:/bld/bld/rattler-build_numpy_1766373812/work/dist/numpy-2.4.0-cp311-cp311-win_amd64.whl#sha256=73a288d2179f5f1426c99bf90b3cc81232ae8e94579705bb5e5219a716553cd8
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optree @ file:///D:/bld/optree_1763124260382/work
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///D:/bld/pandas_1764614503455/work
pillow @ file:///D:/bld/bld/rattler-build_pillow_1767353195/work
posthog==5.4.0
propcache @ file:///D:/bld/propcache_1744524926447/work
protobuf==6.33.2
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pybind11 @ file:///D:/bld/bld/rattler-build_pybind11_1755953378/work
pybind11-global @ file:///D:/bld/bld/rattler-build_pybind11-global_1755953378/work
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks @ file:///D:/bld/pysocks_1733217287171/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.2.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///D:/bld/pyyaml_1758891577448/work
referencing==0.37.0
regex @ file:///D:/bld/regex_1762506879487/work
requests @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_requests_1766926974/work
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors @ file:///D:/bld/safetensors_1763569562911/work
scikit-learn @ file:///D:/bld/bld/rattler-build_scikit-learn_1765801251/work/dist/scikit_learn-1.8.0-cp311-cp311-win_amd64.whl#sha256=6b709d3d1e68d89c77b77f0fb7aa31e6f68e5148d615ecd56f94da3e1262056b
scipy @ file:///C:/bld/scipy-split_1766107338578/work/dist/scipy-1.16.3-cp311-cp311-win_amd64.whl#sha256=4c1a5c75c2b5c1c59e4a27070d8b598350e4e3af66e17d4c4de299c9b4e9fe16
sentence-transformers @ file:///home/conda/feedstock_root/build_artifacts/sentence-transformers_1765540995468/work
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
streamlit==1.52.2
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946057943/work
tenacity==9.1.2
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tokenizers @ file:///D:/bld/tokenizers_1764694778332/work/bindings/python
toml==0.10.2
torch @ file:///C:/bld/libtorch_1764705344950/work
tornado==6.5.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1764088833189/work
typer==0.21.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1765719872007/work
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
win_inet_pton @ file:///D:/bld/win_inet_pton_1733130564612/work
xxhash @ file:///D:/bld/python-xxhash_1762516226414/work
yarl @ file:///D:/bld/yarl_1761336817777/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_zipp_1764460141/work
zstandard==0.25.0
